Desktop/VAI-DE-PIX-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.9.2
pytest==7.4.3
httpx==0.25.2
bcrypt==4.1.2
alembic==1.13.1
email-validator==2.3.0
requests==2.32.5
slowapi==0.1.9
waitress==2.1.2
bleach==6.1.0
apscheduler==3.10.4
